numpy doesnt have np.int any longer

+++ rsl_rl/rsl_rl/datasets/motion_loader.py
@@ -252,8 +252,8 @@ class AMPLoader:
         """Returns frame for the given trajectory at the specified time."""
         p = times / self.trajectory_lens[traj_idxs]
         n = self.trajectory_num_frames[traj_idxs]
-        idx_low, idx_high = np.floor(p * n).astype(np.int), np.ceil(p * n).astype(
-            np.int
+        idx_low, idx_high = np.floor(p * n).astype(int), np.ceil(p * n).astype(
+            int
         )
         all_frame_starts = torch.zeros(
             len(traj_idxs), self.observation_dim, device=self.device
@@ -284,8 +284,8 @@ class AMPLoader:
     def get_full_frame_at_time_batch(self, traj_idxs, times):
         p = (times / self.trajectory_lens[traj_idxs]) % 1
         n = self.trajectory_num_frames[traj_idxs]
-        idx_low, idx_high = np.floor(p * n).astype(np.int), np.ceil(p * n).astype(
-            np.int
+        idx_low, idx_high = np.floor(p * n).astype(int), np.ceil(p * n).astype(
+            int
         )
         all_frame_pos_starts = torch.zeros(
             len(traj_idxs), AMPLoader.POS_SIZE, device=self.device
